+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [816](https://github.com/vegaprotocol/data-node/issues/816) - `v2` accounts `api`
 - [801](https://github.com/vegaprotocol/data-node/issues/801) - Remove distinct select from hyper tables 
 - [570](https://github.com/vegaprotocol/data-node/issues/570) - Add pagination for proposals with ordering of open proposals listed first
+- [481](https://github.com/vegaprotocol/data-node/issues/481) - Add pagination for delegations
 
 ### 🐛 Fixes
 - [594](https://github.com/vegaprotocol/data-node/issues/594) - Fix returning proposal rationale